+++ recipes/install_package.rb
@@ -15,9 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-include_recipe 'docker::default'
-
-install_type = node['drone']['install_type']
-include_recipe "drone::install_#{install_type}"
-
-include_recipe 'drone::_service'
+package 'drone' do
+  source node['drone']['package_url'] if node['drone']['package_url']
+  version node['drone']['package_version'] if node['drone']['package_version']
+end